pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vol-scanner"
version = "0.1.0"
description = "Neural Volatility Surface Arbitrage Scanner"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Pablo Williams" }]
dependencies = [
    "numpy",
    "scipy",
    "scikit-learn",
    "torch",
    "pandas",
    "matplotlib",
    "reportlab",
    "pdfplumber",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]
"src/vol_scanner/figures/*" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
